+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KN91R2QA68Z1Z7947MBR12TF"
+        "01KNGMV8YVV7DWPKDTWPN9WK9X"
     )
     r"""The unique identifier of the data source"""