+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KN91R2Q3VNMCGCXSKVHK4SY1"
+        "01KNGMV8YTT4PVQJAXWJ86NE34"
     )
     r"""The unique identifier of the data source"""